+++ events-service/src/main/java/com/event4u/eventsservice/service/LocationService.java
@@ -1,5 +1,8 @@
 package com.event4u.eventsservice.service;
 
+import com.event4u.eventsservice.exceptionHandling.AlreadyExistsException;
+import com.event4u.eventsservice.exceptionHandling.NotFoundException;
+import com.event4u.eventsservice.model.Category;
 import com.event4u.eventsservice.model.Location;
 import com.event4u.eventsservice.repository.LocationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +26,11 @@ public class LocationService {
     }
 
     public Location findById(Long id) {
-        return locationRepository.findById(id).orElseThrow();
+        Optional<Location> l = locationRepository.findById(id);
+        if (!l.isPresent()) {
+            throw new NotFoundException("Location with id " + id.toString());
+        }
+        return l.get();
     }
 
     public Long count() {
@@ -31,19 +38,37 @@ public class LocationService {
     }
 
     public void deleteById(Long locationId) {
+        if (!locationRepository.existsById(locationId)) {
+            throw new NotFoundException("Location with id " + locationId.toString());
+        }
         locationRepository.deleteById(locationId);
     }
 
     public Location addNewLocation(Point coordinates, String city, String country) {
+        List<Location> locations = findAll();
+        locations.forEach(l-> {
+            if (l.getCity().equals(city) && l.getCountry().equals(country)) {
+                throw new AlreadyExistsException("Location " + city + " " + country);
+            }
+        });
         return locationRepository.save(new Location(coordinates, city, country));
     }
 
-    public Optional<Location> updateLocation(Long id, Point coordinates, String city, String country) {
+    public Location updateLocation(Long id, Point coordinates, String city, String country) {
+        if (!locationRepository.existsById(id)) {
+            throw new NotFoundException("Location with id " + id.toString());
+        }
+        List<Location> locations = findAll();
+        locations.forEach(l-> {
+            if (l.getCity().equals(city) && l.getCountry().equals(country)) {
+                throw new AlreadyExistsException("Location " + city + " " + country);
+            }
+        });
         return locationRepository.findById(id).map(location-> {
             location.setCoordinates(coordinates);
             location.setCity(city);
             location.setCountry(country);
             return locationRepository.save(location);
-        });
+        }).orElseThrow();
     }
 }
