+++ events-service/src/main/java/com/event4u/eventsservice/service/UserService.java
@@ -1,5 +1,7 @@
 package com.event4u.eventsservice.service;
 
+import com.event4u.eventsservice.exceptionHandling.AlreadyExistsException;
+import com.event4u.eventsservice.exceptionHandling.NotFoundException;
 import com.event4u.eventsservice.model.Event;
 import com.event4u.eventsservice.model.User;
 import com.event4u.eventsservice.repository.UserRepository;
@@ -7,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -14,7 +17,11 @@ public class UserService {
     private UserRepository userRepository;
 
     public User getUserById(Long id) {
-        return userRepository.findById(id).orElseThrow();
+        Optional<User> u = userRepository.findById(id);
+        if (!u.isPresent()) {
+            throw new NotFoundException("User with id " + id.toString());
+        }
+        return u.get();
     }
 
     public ArrayList<User> getAllUsers() {
@@ -25,10 +32,16 @@ public class UserService {
     }
 
     public User createUser(Long id) {
+        if (userRepository.existsById(id)) {
+            throw new AlreadyExistsException("User with id " + id.toString());
+        }
         return userRepository.save(new User(id));
     }
 
     public void deleteUser(Long id) {
+        if(!userRepository.existsById(id)) {
+            throw new NotFoundException("User with id " + id.toString());
+        }
         userRepository.deleteById(id);
     }
